[DELETE] Draggable.cs
- 기존 Draggable 스크립트 삭제

- SimpleDraggable -> Draggable 명칭 변경
