--- Assets/Scripts/Tower/SimpleDraggable.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4c00abfbb15a98241a86ff16bd684321
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
